requirements.txt
anyio==3.6.1
asyncpg==0.26.0
autopep8==1.7.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
databases==0.6.1
fastapi==0.85.0
greenlet==1.1.3
h11==0.14.0
httptools==0.5.0
idna==3.4
ormar==0.11.3
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
python-dotenv==0.21.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
toml==0.10.2
typing_extensions==4.3.0
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
